Registered most Mezzanine models for translation.
With the exception of Keyword and Setting as they need some special
handling. Including the abstract Displayable as it is instantiated
for sitemap generation.

Base classes need to be registered for translation before their
descendants, so the installed apps order needed to be modifed a bit.

Having a test pointing out unregistered textual fields should make
maintaining the integration easier -- most such new fields should be
made translatable (add them to the appropriate translation.py).