Let TranslationAdmin be used for anything that potentially can have translated fields. Use modeltranslation scripting to display translation fields as tabs on default.
Compared to the previous version we can now rely on the jQuery UI bundled with
Mezzanine. Hurray! :-) Now there's also a ``JQUERY_UI_FILENAME`` setting, making
the additions to the base template fairly general; one thing left requiring a line
of fancy code is linking to the UI theme.

Conflicts:

	mezzanine/core/templates/admin/base_site.html