Patched processing of fields derived from translated fields.
Fields that need to be generated for all languages are:
* Slugged.slug,
* MetaData.description,
* Displayable.short_url,
* Page.titles.
Moreover, when generating description one needs to be careful not to
use existing translated description.

Factored out two helpers that may be useful for similar tasks in user
code. The disable_fallbacks() context prevents getting value from a
default language when one for the active one is missing, useful when
you need to know if a value in the current language is available.
The for_all_languages() function executes a function with each of the
available languages activated in turn.

The for_all_languages() function and the disable_fallbacks() manager
are used outside of USE_MODELTRANSLATION conditions, so tests with
enabled and disabled content translation are needed for them.
Model.save() should not use any more queries with content translation
than without it, query counts in both save performance tests should be
updated in unison.

TODO: Propose light changes to trans_null upstream.
TODO: Investigate Page.save() queries count on master.