Update all translations of derived / generated fields when models are saved (Slugged.slug, MetaData.description and Page.titles).
Helpers for executing code without language fallbacks (when you
need to know when a value in the current language is available)
and for iterating a function with each of available languages
activated in turn.

Slug handling is a bit cleaner than in previous version, but still
it's one of the more involved pieces of the modification. Automatic
translated slugs are too nice to drop it however.

Conflicts:

	mezzanine/core/models.py