Automated modeltranslation database setup; added a resavemodels command.
During createdb sync_translation_fields, update_translation_fields and
resavemodels are executed, giving an out-of-the-box internationalized
project. It is most convenient to configure languages and enable
content translation in settings before running createdb.

With some fields, one cannot always rely on translation fallbacks, as
they don't work with filtering / lookups. For example, if you create a
new project, manually migrate and update translation fields, you won't
be able to browse most pages in any non-default language until you give
them a chance to generate translated slugs.

The new command is an opt-in, to apply for it add resavemodels_command
to the models class. By default, with modeltranslation enabled, Slugged,
MetaData and Page (and their subclasses) are resaved with the command.

TODO: The post_syncdb infrastructure is risky (the signal is outdated,
      moreover the initial setup should only be done after the inital
      migration). Needs changes on master.