Implemented keywords translation.
Assignments are intended to be the same for every language, thus only
Keyword.title and keywords_string (on concrete subclasses of MetaData)
are made translatable, with the latter recomputed on change for all
languages as with other derived attributes.

Page admin only allows entering keywords in the active language (there
doesn't seem to be any sensible way to inline translations), so
KeywordAdmin (under Content menu) is exposed separately to make it
possible to translate keywords.