Made modeltranslation fields on lists narrower, so a few translations can fit. This also gives a cleaner layout without modeltranslation (e.g. cartridge sale inline).
It would be nice to display just one translation for a whole inline
and control which one with the main language switch / tabs. However,
this would require a specific script for Mezzanine's dynamic, sortable
inlines. There's a sketch of a solution on the branch:
modeltranslation/feature/hide-inactive-untabbed.

Note: This is the last of the cherry-picked commits from the old branch,
what follows (if anything ;-) is new.