Made modeltranslation fields on lists narrower, so a couple of translations can fit.
This also gives a cleaner layout with vanilla admin in some places
(e.g. cartridge sale inline).

It would be nice to display just one translation for a whole inline
and control which one is displayed with the main language switch / tabs.
However, this would require a specific script for Mezzanine's dynamic,
sortable inlines. There's a sketch of a solution on the branch:
modeltranslation/feature/hide-inactive-untabbed.