Fixed testing apps not in INSTALLED_APPS that need to have their default settings loaded.
This is probably a fairly common case (e.g. new project without
uncommenting mezzanine.accounts), to which there doesn't seem to
be a straightforward solution.

The main issue is that it's not possible to get apps that will be
tested from the test suite, before they are imported.
Django relies on the unittest loader to do the modules discovery,
which tries importing and if it fails (say due to missing settings)
only leaves a scarce info in the test suite (a pseudo-TestCase with
just an attribute key referring to what it was trying to import).
Moreover even if an import succeeds we still can't be sure if some
tests won't fail due to skipping the defaults registration.

The proper solution would probably be adding some before-import
hooks to the unittest.loader.TestLoader discover and loadTestFromName
methods.

Instead we find and process all "defaults.py" in the lowest-level
package for each test label and all its subpackages (the latter is
needed with labels like "mezzanine").