Removed a call private set_installed_apps and the whole manual pre-discovery at the cost of making the tests importable without the Mezzanine setup.
The "import defaults" statements in tests are needed when testing
non-installed apps -- the test runner will attempt an import during
test discovery, before a list of what is to be tested is known.

Other than the just removed "pre-discovery", it would be possible
to solve this overriding unittest.TestLoaded (but a lot needs to
be overrided) or using sys.path_hooks (but this seems like an
overshot for such a case).