Added modeltranslation integration settings tests.
The main concern here is to check, if introducing the new code does
not cause any side-effects when USE_MODELTRANSLATION is false. Any
additional tests for that should go into the NoContentTranslation
test case.

This commit also introduces a TestRunner that automatically adds
translation columns to the test database.
The test runner relies on some proposed modeltranslation changes:
https://github.com/deschler/django-modeltranslation/pull/276.